milestone 1: containers working

+++ init_database.sql
@@ -1,3 +1,4 @@
+-- Database Initialization Script --
 
 -- Source Columns: region,origin_coord,destination_coord,datetime,datasource
 
@@ -16,8 +17,13 @@ CREATE TABLE IF NOT EXISTS trip_fact
     region VARCHAR(255) NOT NULL,
     datasource VARCHAR(255) NOT NULL,
     tripdate TIMESTAMP NOT NULL,
-    lat1 double NOT NULL,
-    lng1 double NOT NULL,
-    lat2 double NOT NULL,
-    lng2 double NOT NULL
+    lat1 float NOT NULL,
+    lng1 float NOT NULL,
+    lat2 float NOT NULL,
+    lng2 float NOT NULL
 );
+
+-- select * from trip_fact
+-- select * from region_dimension
+
+-- End of Script --
